+++ config/releases.exs
@@ -22,6 +22,11 @@ config :td_bg, TdBg.Scheduler,
   jobs: [
     [
       schedule: "@reboot",
+      task: {TdBg.Jobs.UpdateDomainFields, :run, []},
+      run_strategy: Quantum.RunStrategy.Local
+    ],
+    [
+      schedule: "@reboot",
       task: {TdCache.CacheCleaner, :clean, [["domains:root", "domains:id_to_parent_ids"]]},
       run_strategy: Quantum.RunStrategy.Local
     ],